Add convenience functions and documentation of node ordering
